refactor(reservation-persistence): JpaSeatPersistenceAdapter 패키지 분리
